+++ packages/brick_offline_first_with_graphql/test/graphql_offline_request_queue_test.dart
@@ -2,7 +2,6 @@ import 'package:brick_offline_first_with_graphql/src/graphql_offline_queue_link.
 import 'package:brick_offline_first_with_graphql/src/graphql_offline_request_queue.dart';
 import 'package:brick_offline_first_with_graphql/src/graphql_request_sqlite_cache_manager.dart';
 import 'package:test/test.dart';
-import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
 void main() {