+++ packages/brick_offline_first_with_graphql/test/graphql_offline_sqlite_cache_manager_test.dart
@@ -5,7 +5,6 @@ import 'package:test/test.dart';
 import 'package:gql_exec/gql_exec.dart';
 import 'package:gql/language.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
-import 'package:sqflite_common/sqlite_api.dart';
 
 import '__helpers__.dart';