+++ packages/brick_offline_first_with_rest/test/offline_queue/rest_request_sqlite_cache_test.dart
@@ -5,7 +5,6 @@ import 'package:http/http.dart' as http;
 import 'package:mockito/mockito.dart';
 import 'package:logging/logging.dart';
 import 'package:brick_offline_first/src/offline_queue/request_sqlite_cache.dart';
-import 'package:sqflite_common/sqlite_api.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
 class MockLogger extends Mock implements Logger {}