+++ packages/brick_rest_generators/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+## 3.0.1
+
+* Support Dart 3
+
 ## 3.0.0
 
 * Update reference from `brick_rest/rest.dart` to `brick_rest/brick_rest.dart`